+++ terraform/environment/main.tf
@@ -1,0 +1,11 @@
+variable "environment" {
+  type = string
+}
+
+variable "root_domain" {
+  type = string
+}
+
+variable "inventory_file" {
+  type = string
+}