+++ terraform/modules/aws-network-load-balancer/outputs.tf
@@ -1,0 +1,3 @@
+output "fqdn" {
+  value = aws_lb.nlb.dns_name
+}