+++ terraform/modules/aws-vpc-security-groups/variables.tf
@@ -3,8 +3,3 @@ variable "vpc_id" {
   description = "ID of VPC these security groups are for."
 }
 
-variable "s3_CIDRs" {
-  type        = list(string)
-  description = "subnets that S3 gateways we are using exist in."
-}
-