+++ terraform/modules/aws-vpc/outputs.tf
@@ -10,3 +10,6 @@ output "private_subnets" {
   value = module.vpc.private_subnets
 }
 
+output "private_route_table_ids" {
+  value = module.vpc.private_route_table_ids
+}